+++ packages/react-static/src/bootstrapApp.js
@@ -1,20 +1,12 @@
 /* eslint-disable import/no-dynamic-require */
 import * as React from 'react'
 import { staticInfoContext } from './browser/hooks/useStaticInfo'
-
-const OriginalSuspense = React.Suspense
-
-function Suspense({ key, children, ...rest }) {
-  return typeof document !== 'undefined' ? (
-    <OriginalSuspense key={key} {...rest}>
-      {children}
-    </OriginalSuspense>
-  ) : (
-    <React.Fragment key={key}>{children}</React.Fragment>
-  )
-}
+import Suspense from './Suspense'
 
 // Override the suspense module to be our own
+// This is expected to break when using preact
+// In order to make it work with preact 10, use `patch-package` to remove those 2 lines
+// Reference: https://github.com/react-static/react-static/pull/1500
 React.Suspense = Suspense
 React.default.Suspense = Suspense